requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.8.2
pandas
openpyxl
requests
markdown
# Query Understanding — Component 1: Preprocessing
nltk
unidecode
pyspellchecker
# Query Understanding — Budget Classifier (ONNX)
scikit-learn
skl2onnx
onnxruntime